+++ src/main/java/sw_10/p3_backend/Logic/TechnicianLogic.java
@@ -1,19 +1,25 @@
 package sw_10.p3_backend.Logic;
 
 import org.springframework.stereotype.Service;
+import sw_10.p3_backend.Model.ResourceOrder;
 import sw_10.p3_backend.Model.Technician;
 import sw_10.p3_backend.Repository.TechnicianRepository;
+import sw_10.p3_backend.Logic.ResourceOrderLogic;
 import sw_10.p3_backend.exception.InputInvalidException;
 import sw_10.p3_backend.exception.NotFoundException;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class TechnicianLogic {
     private final TechnicianRepository technicianRepository;
+    private final ResourceOrderLogic resourceOrderLogic;
 
-    public TechnicianLogic(TechnicianRepository technicianRepository) {
+    public TechnicianLogic(TechnicianRepository technicianRepository, ResourceOrderLogic resourceOrderLogic) {
         this.technicianRepository = technicianRepository;
+        this.resourceOrderLogic = resourceOrderLogic;
     }
 
     public Technician TechnicianByType(String type) {
@@ -74,11 +80,23 @@ public class TechnicianLogic {
         try {
             Technician technician = technicianRepository.findByType(type);
             if (technician != null) {
+                /*if (technician.getWorkHours() > 0) {
+                    throw new InputInvalidException("Technician has " + technician.getWorkHours() + " work hours assigned. Cannot delete technician with work");
+                }*/
+                List<ResourceOrder> resourceOrders = resourceOrderLogic.findResourceName(technician.getType());
+                if (!resourceOrders.isEmpty()) {
+                    throw new InputInvalidException("Cannot delete technician with assigned BladeTasks. Currently assigned " + resourceOrders.size() + " blade tasks");
+
+                }
+                // TODO ressourceoOrders check efter typex
+                // kald fra TechnicianLogic til ressourceOrderLogic som kalder resourceorderrepository
                 technicianRepository.delete(technician);
                 return technician;
             } else {
                 throw new NotFoundException("Technician not found with type: " + type);
             }
+        } catch (InputInvalidException e) {
+            throw new InputInvalidException(e.getMessage());
         } catch (NotFoundException e) {
             throw new NotFoundException(e.getMessage());
         } catch (Exception e) {
